Refactor app.tsx to break out all keymap functionality into custom hook
